development

+++ src/map.h
@@ -14,8 +14,8 @@ public:
 	struct single_landmark_s{
 
 		int id_i ; // Landmark ID
-		float x_f; // Landmark x-position in the map (global coordinates)
-		float y_f; // Landmark y-position in the map (global coordinates)
+		double x_f; // Landmark x-position in the map (global coordinates)
+		double y_f; // Landmark y-position in the map (global coordinates)
 	};
 
 	std::vector<single_landmark_s> landmark_list ; // List of landmarks in the map
